requirements.txt
# Core
python-dotenv==1.0.1
pydantic==2.10.4
requests==2.32.5

# DuckDB + data
duckdb==1.1.3
pandas==2.3.3
numpy==2.4.0
tabulate==0.9.0

# LangChain + LangGraph (0.3.x compatible family)
langchain==0.3.14
langchain-core==0.3.29
langchain-text-splitters==0.3.5
langchain-community==0.3.14
langgraph==0.2.61
langsmith==0.2.11

# Gemini (keep only if you still want Gemini option)
google-genai==0.7.0
langchain-google-genai==2.0.9

# Vector store for schema RAG
chromadb==0.5.23
sentence-transformers==5.2.0